# docs/swagger.yaml
basePath: /kitchencontrol/api/v1
definitions:
...
  dto.CreateProductDTO:
    properties:
      category:
        type: string
      description:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
...
  dto.ProductDTO:
    properties:
      category:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is fiap ddd challenge project.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fiap Challenge DDD API
  version: "1.0"
paths:
...
  /products:
    get:
      consumes:
      - application/json
      description: Get details of all products
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ProductDTO'
            type: array
        "400":
          description: Invalida data
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: No products foundr
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Create Product in DB
      parameters:
      - description: cria novo produto
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProductDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProductDTO'
        "400":
          description: invalid data
          schema:
            type: string
        "409":
          description: product already exists
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Create Product
      tags:
      - Products
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: Delete details of a Product by id
      parameters:
      - description: Product id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProductDTO'
        "400":
          description: Invalid data
          schema:
            type: string
      summary: Delete a Product
      tags:
      - Products
    get:
      consumes:
      - application/json
      description: Get details of a Product by id
      parameters:
      - description: Product id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProductDTO'
        "404":
          description: Product not found
          schema:
            type: string
      summary: Get a Product
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Update details of a Product by id
      parameters:
      - description: Product id
        in: path
        name: id
        required: true
        type: string
      - description: Product data
        in: body
        name: Product
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProductDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProductDTO'
        "400":
          description: Invalid data
          schema:
            type: string
        "404":
          description: Product not found
          schema:
            type: string
      summary: Update a Product
      tags:
      - Products
  /products/category/{id}:
    get:
      consumes:
      - application/json
      description: Get details of a Product by category
      parameters:
      - description: Product category
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProductDTO'
        "404":
          description: Product not found
          schema:
            type: string
        "500":
          description: Product not found
          schema:
            type: string
      summary: Get a Product by category
      tags:
      - Products
swagger: "2.0"